requirements.txt
beautifulsoup4==4.12.3
boto3==1.36.8
botocore==1.36.8
gTTS==2.5.4
moviepy==2.2.1
requests==2.32.3
rich==13.9.4
toml==0.10.2
tomlkit==0.13.2
Flask==3.1.1
clean-text==0.6.0
unidecode==1.4.0
ffmpeg-python==0.2.0
elevenlabs==1.57.0
yt-dlp==2025.10.22
Pillow>=10.0.0
pyttsx3==2.98
tqdm>=4.66.0
